+++ NEWS.md
@@ -1,6 +1,6 @@
 # Keng 2024.12.15
 
-* `power_r()` has been added to conduct post-hoc and prior power analysis, and plan the sample size for Pearson's r.
+* `power_r()` has been added to conduct post-hoc and a priori power analysis, and plan the sample size for Pearson's r.
 * `power_lm()` has reorganized its output.
 * `power_r()` and `power_lm()` return the `Keng_power` class, which has `print()` and `plot()` methods.
 
@@ -9,7 +9,7 @@
 * `test_r()` has added Fisher's z.
 * `compare_lm()` has added f_squared and post-hoc power.
 * `calc_PRE()` has been added to calculate PRE from partial correlation, Cohen's f, or f_squared.
-* `power_lm()` has been added to conduct post-hoc and prior power analysis, and plan the sample size for one or a set of predictors in linear regression.
+* `power_lm()` has been added to conduct post-hoc and a priori power analysis, and plan the sample size for one or a set of predictors in linear regression.
 
 # Keng 2024.11.17